+++ src/briefy/leica/views/dashboard/finance.py
@@ -1,5 +1,6 @@
 """Views to handle Finance Dashboards."""
 from briefy.leica.models.dashboard.finance import DashboardFinanceOrder
+from briefy.leica.views.dashboard import DELIVERED_ORDERS_COLS
 from briefy.leica.views.dashboard import ORDER_PROJECT_COLS
 from briefy.ws import CORS_POLICY
 from briefy.ws.resources import SQLQueryService
@@ -85,6 +86,89 @@ class DashboardFinanceOrderService(SQLQueryService):
     (SELECT i.id, i.state, i.title
     FROM items as i JOIN projects as p on i.id = p.id
     WHERE i.state = 'ongoing') as projects
+    on orders.project_id = projects.id
+
+    ) as active_orders GROUP BY
+    active_orders.title,
+    active_orders.project_id
+    ORDER BY active_orders.title
+    """
+
+    def query_params(self, query: str) -> str:
+        """Apply query parameters based on request.
+
+        This is supposed to be specialized by resource classes.
+
+        :query: string with a query to be parametrized
+        :returns: string with a query after adding parameters
+        """
+        principal_id = self.request.user.id
+        return query.format(principal_id=principal_id, type='order')
+
+    def transform(self, data: list) -> list:
+        """Transform data items after query execution.
+
+        :data: list of records to be transformed
+        :returns: list of records after transformation
+        """
+        for item in data:
+            project_id = item.pop('project_id')
+            item['absolute_url'] = f'/projects/{project_id}'
+        return data
+
+
+COLLECTION_PATH = '/dashboards/finance/delivered'
+PATH = COLLECTION_PATH + '/{id}'
+
+
+@resource(collection_path=COLLECTION_PATH,
+          path=PATH,
+          cors_policy=CORS_POLICY,
+          factory=DashboardFinanceFactory)
+class DashboardFinanceDeliveredOrdersService(SQLQueryService):
+    """Dashboard Finance: Delivered Orders Service."""
+
+    _columns_map = DELIVERED_ORDERS_COLS
+
+    _collection_query = """
+    SELECT
+    count(active_orders.id) as total,
+    active_orders.title,
+    active_orders.project_id,
+    sum(
+    CASE WHEN
+    active_orders.state = 'delivered'
+    AND active_orders.accept_date IS NULL
+    THEN 1 ELSE 0
+    END) as newly_delivered,
+    sum(
+    CASE WHEN
+    active_orders.state IN ('refused', 'in_qa')
+    AND active_orders.accept_date IS NOT NULL
+    THEN 1 ELSE 0
+    END) as further_revision,
+    sum(
+    CASE WHEN active_orders.state = 'delivered'
+    AND active_orders.accept_date IS NOT NULL
+    THEN 1 ELSE 0
+    END) as re_delivered,
+    sum(
+    CASE WHEN active_orders.state = 'accepted'
+    THEN 1 ELSE 0
+    END) as completed
+    FROM
+
+    (SELECT DISTINCT orders.id, orders.project_id,
+    projects.title, orders.state, orders.accept_date FROM
+
+    (SELECT i.id, i.state, i.title, o.accept_date, o.project_id
+    FROM items as i JOIN orders as o on i.id = o.id
+    WHERE i.type = '{type}' AND
+    i.state IN ('delivered', 'accepted', 'in_qa', 'refused')
+    ) as orders JOIN
+
+    (SELECT i.id, i.state, i.title
+    FROM items as i JOIN projects as p on i.id = p.id) as projects
     on orders.project_id = projects.id
 
     ) as active_orders GROUP BY